+++ test/common/data-registry.ts
@@ -1,3 +1,3 @@
-import { RegistryUrl } from "../../../src/domain/registry-url";
+import { RegistryUrl } from "../../src/domain/registry-url";
 
 export const exampleRegistryUrl = RegistryUrl.parse("https://example.com");
