+++ test/integration/app/get-registry-packument-version.test.ts
@@ -4,8 +4,8 @@ import { DomainName } from "../../../src/domain/domain-name";
 import { Registry } from "../../../src/domain/registry";
 import { SemanticVersion } from "../../../src/domain/semantic-version";
 import { GetRegistryPackument } from "../../../src/io/packument-io";
-import { buildPackument } from "../../unit/domain/data-packument";
-import { exampleRegistryUrl } from "../../unit/domain/data-registry";
+import { buildPackument } from "../../common/data-packument";
+import { exampleRegistryUrl } from "../../common/data-registry";
 import { mockFunctionOfType } from "./func.mock";
 
 describe("fetch registrypackument version", () => {
