+++ test/integration/app/login.test.ts
@@ -3,7 +3,7 @@ import { StoreNpmAuthToken } from "../../../src/app/put-npm-auth-token";
 import { PutRegistryAuth } from "../../../src/app/put-registry-auth";
 import { GetAuthToken } from "../../../src/io/get-auth-token";
 import { noopLogger } from "../../../src/logging";
-import { exampleRegistryUrl } from "../../unit/domain/data-registry";
+import { exampleRegistryUrl } from "../../common/data-registry";
 import { mockFunctionOfType } from "./func.mock";
 
 const exampleUser = "user";
