+++ test/unit/cli/dependency-logging.test.ts
@@ -6,7 +6,7 @@ import {
 } from "../../../src/domain/dependency-graph";
 import { stringifyDependencyGraph } from "../../../src/cli/dependency-logging";
 import { makePackageReference } from "../../../src/domain/package-reference";
-import { exampleRegistryUrl } from "../domain/data-registry";
+import { exampleRegistryUrl } from "../../common/data-registry";
 import { PackumentNotFoundError } from "../../../src/common-errors";
 import { unityRegistryUrl } from "../../../src/domain/registry-url";
 import { VersionNotFoundError } from "../../../src/domain/packument";
